refactor(mm-next/story): move wide component `nav-subtitle-navigator` into shared
