+++ packages/mirror-media-next/components/story/wide/index.js
@@ -11,7 +11,7 @@ import DonateLink from '../shared/donate-link'
 import HeroImageAndVideo from '../shared/hero-image-and-video'
 import Credits from '../shared/credits'
 import DonateBanner from '../shared/donate-banner'
-import NavSubtitleNavigator from './nav-subtitle-navigator'
+import NavSubtitleNavigator from '../shared/nav-subtitle-navigator'
 import MoreInfoAndTag from '../shared/more-info-and-tag'
 import Date from '../shared/date'
 import ButtonCopyLink from '../shared/button-copy-link'
